Cleanly separate HTIF code; don't poll keyboard on timer interrupt
