+++ machine/machine.mk.in
@@ -18,6 +18,7 @@ machine_hdrs = \
 machine_c_srcs = \
   mtrap.c \
   minit.c \
+  htif.c \
   emulation.c \
   muldiv_emulation.c \
   fp_emulation.c \
