+++ machine/mcall.h
@@ -4,7 +4,6 @@
 #define MCALL_HART_ID 0
 #define MCALL_CONSOLE_PUTCHAR 1
 #define MCALL_CONSOLE_GETCHAR 2
-#define MCALL_HTIF_SYSCALL 3
 #define MCALL_SEND_IPI 4
 #define MCALL_CLEAR_IPI 5
 #define MCALL_SHUTDOWN 6
