+++ pk/frontend.c
@@ -4,8 +4,8 @@
 #include "atomic.h"
 #include "frontend.h"
 #include "sbi.h"
-#include "mcall.h"
 #include "syscall.h"
+#include "htif.h"
 #include <stdint.h>
 
 long frontend_syscall(long n, uint64_t a0, uint64_t a1, uint64_t a2, uint64_t a3, uint64_t a4, uint64_t a5, uint64_t a6)
@@ -24,7 +24,7 @@ long frontend_syscall(long n, uint64_t a0, uint64_t a1, uint64_t a2, uint64_t a3
   magic_mem[6] = a5;
   magic_mem[7] = a6;
 
-  do_mcall(MCALL_HTIF_SYSCALL, magic_mem);
+  htif_syscall((uintptr_t)magic_mem);
 
   long ret = magic_mem[0];
